Improvement: Data structure for Edge is better. Improvement: The code doesn't generate duplicate edges and just overwrite the cost.

+++ OptimiseRoute/PathEdge.h
@@ -11,10 +11,9 @@
 @class PathNode;
 
 @interface PathEdge : NSObject
-    @property (assign) NSUInteger cost;
+    @property (assign) CGFloat cost;
     @property (nonatomic, weak) PathNode *nodeOne;
     @property (nonatomic, weak) PathNode *nodeTwo;
-
 -(void) printEdge;
 
 @end // PathEdge
